--- client/src/components/modal/modal.js
@@ -1,15 +1,0 @@
-import React from "react";
-import Login from "../auth/Login";
-import Register from "../auth/Register";
-import { Mutation } from "react-apollo";
-
-class Modal extends React.Component {
-  constructor(props) {
-    super(props);
-    this.handleClick = this.handleClick.bind(this);
-    // this.state = { closing: false };
-  }
-
-}
-
-export default Modal;
