+++ client/src/components/nav/navbar.css
@@ -1,19 +1,25 @@
-;.header {
+.header nav {
   position: fixed;
   left: 0;
   top: 0;
   width: 100%;
-  height: 3.5.rem;
+  height: 3.5rem;
   padding: 0 1rem;
   display: flex;
   align-items: center;
+  justify-content: space-between
 }
 
 .header :hover {
   background: lightgrey;
 }
 
-.header-logo h1 {
+.header-logo {
+  display: flex;
+  background-position: center;
+  background-size: cover;
+  height:40px;
+  width: 40px;
   margin: 0;
   font-size: 1.5rem;
 }
